Core - SpaceShip can now collide, Road now has a list of its SpaceShips | Geom floating type float -> double | Refactor more ArrayList converted into List

+++ src/net/neczpal/interstellarwar/desktop/geom/Point.java
@@ -3,7 +3,7 @@ package net.neczpal.interstellarwar.desktop.geom;
 import static org.lwjgl.opengl.GL11.*;
 
 public class Point {
-	float x, y;
+	double x, y;
 
 	/**
 	 * Erstellt ein Punkt
@@ -11,7 +11,7 @@ public class Point {
 	 * @param x Die Position der X-Achse
 	 * @param y Die Position der Y-Achse
 	 */
-	public Point (float x, float y) {
+	public Point (double x, double y) {
 		setPosition (x, y);
 	}
 
@@ -32,7 +32,7 @@ public class Point {
 	 * @param dx X-Entfernung
 	 * @param dy Y-Entfernung
 	 */
-	public void move (float dx, float dy) {
+	public void move (double dx, double dy) {
 		x += dx;
 		y += dy;
 	}
@@ -43,43 +43,43 @@ public class Point {
 	 * @param p     Der Punkt
 	 * @param angle Der Winkel
 	 */
-	public void rotate (Point p, float angle) {
-		float lx = p.x - x;
-		float ly = p.y - y;
-		float cos = (float) Math.cos (angle);
-		float sin = (float) Math.sin (angle);
+	public void rotate (Point p, double angle) {
+		double lx = p.x - x;
+		double ly = p.y - y;
+		double cos = Math.cos (angle);
+		double sin = Math.sin (angle);
 
-		x = p.x - (int) (lx * cos - ly * sin);
-		y = p.y - (int) (lx * sin + ly * cos);
+		x = p.x - (lx * cos - ly * sin);
+		y = p.y - (lx * sin + ly * cos);
 	}
 
 	/**
 	 * @param p Der Punkt
 	 * @return Die Entfernung von 'p' Punkt
 	 */
-	public float distance (Point p) {
-		return (float) Math.sqrt (Math.pow (p.x - x, 2) + Math.pow (p.y - y, 2));
+	public double distance (Point p) {
+		return Math.sqrt (Math.pow (p.x - x, 2) + Math.pow (p.y - y, 2));
 	}
 
 	//GETTERS, SETTERS
 
-	public float getX () {
+	public double getX () {
 		return x;
 	}
 
-	public void setX (float x) {
+	public void setX (double x) {
 		this.x = x;
 	}
 
-	public float getY () {
+	public double getY () {
 		return y;
 	}
 
-	public void setY (float y) {
+	public void setY (double y) {
 		this.y = y;
 	}
 
-	public final void setPosition (float x, float y) {
+	public final void setPosition (double x, double y) {
 		setX (x);
 		setY (y);
 	}
